Sound api improvements, rockboy sound, contributed by xshock. Playback of sound currently only works in boost mode, needs fixing.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@6226 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/rockboy/pcm.h
@@ -9,7 +9,7 @@ struct pcm
 {
 	int hz, len;
 	int stereo;
-	byte *buf;
+	short *buf;
 	int pos;
 };
